requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohttp==3.9.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Babel==2.14.0
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.12.1
celery==5.3.6
celery-types==0.23.0
certifi==2023.11.17
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorclass==2.2.2
constantly==23.10.4
convertdate==2.4.0
cryptography==42.0.5
dadata==21.10.1
daphne==4.1.2
dill==0.4.0
Django==5.0.6
django-bootstrap5==24.3
django-ckeditor-5==0.2.12
django-debug-toolbar==4.3.0
django-dynamic-formsets==0.0.8
django-redis==5.4.0
django-widget-tweaks==1.5.0
django-xff==1.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.16.7
emoji==2.14.0
et-xmlfile==1.1.0
flower==2.0.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
hijri-converter==2.3.1
htmldocx==0.0.6
httpcore==1.0.2
httpx==0.26.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jdcal==1.4.1
Jinja2==3.1.3
kombu==5.3.4
korean-lunar-calendar==0.3.1
loguru==0.7.2
lxml==5.1.0
magic-filter==1.0.12
MarkupSafe==2.1.3
msgpack==1.0.8
msgpack-python==0.5.6
msoffice2pdf==0.0.8
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.26.0
numpy==2.1.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pandas-stubs==2.3.0.250703
pathspec==0.12.1
pillow==10.4.0
pip-upgrader==1.4.15
platformdirs==4.1.0
plotly==6.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.11.10
pydantic_core==2.33.2
PyJWT==2.10.1
PyMeeus==0.5.12
pyOpenSSL==24.1.0
pypng==0.20220715.0
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.1.0
python-magic==0.4.27
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.4
redis-cli==1.0.1
requests==2.31.0
rfc3986==2.0.0
service-identity==24.1.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
telebot==0.0.5
terminaltables==3.1.10
tornado==6.4
tqdm==4.66.1
Twisted==24.3.0
txaio==23.1.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.43.0
xlrd==2.0.1
yarl==1.9.4
zope.interface==6.2
